requirements.txt
# torch
# torchvision
# timm
# transformers

# opencv-python
# numpy
# Pillow
# scikit-learn

# facenet-pytorch

# flask

flask
torch
torchvision
timm
opencv-python
numpy
Pillow
retina-face
